x_uploader-revision
Pulled upload_image_to_twitter function out of the
