+++ MLB_Matchup/src/twitter_image_generator.py
@@ -188,6 +188,14 @@ class TwitterImageGenerator:
         draw.text((x, 130), venue, font=self.font_mid, fill=self.text_color)
         x, y = self.get_centered_text_xy(draw, location, self.font_mid, (400, 40))
         draw.text((x, 160), location, font=self.font_mid, fill=self.text_color)
+
+        # Draw team records - centered below abbreviations
+        # away_record = game_data.get('away_record', 'TBD')
+        # home_record = game_data.get('home_record', 'TBD')
+        # x, y = self.get_centered_text_xy(draw, away_record, self.font_mid, (400, 40))
+        # draw.text((x, 200), away_record, font=self.font_mid, fill=self.text_color)
+        # x, y = self.get_centered_text_xy(draw, home_record, self.font_mid, (400, 40))
+        # draw.text((x, 230), home_record, font=self.font_mid, fill=self.text_color)
         
         # Draw lineups - just the text, no table drawing
         away_lineup = game_data.get('away_lineup', [])
